[fix/UnitMapping.vue] 新增共用組件
